// src/utils/types.ts
type DAOMembership = 'token' | 'multisig';
type ProposalCreationEligibility = 'token' | 'anyone' | 'multisig';
export type CreateDaoFormData = {
  blockchain: {
    id: number;
    label: string;
    network: string;
  };
  daoLogo: Blob;
  daoName: string;
  daoEnsName: string;
  daoSummary: string;
  tokenName: string;
  tokenSymbol: string;
  tokenDecimals: number;
  tokenTotalSupply: number;
  tokenTotalHolders: number | undefined;
  tokenType: TokenType;
  isCustomToken: boolean;
  links: {name: string; url: string}[];
  wallets: TokenVotingWalletField[];
  tokenAddress: InputValue;
  durationMinutes: string;
  durationHours: string;
  durationDays: string;
  minimumApproval: string;
  minimumParticipation: string;
  eligibilityType: ProposalCreationEligibility;
  eligibilityTokenAmount: number | string;
  support: string;
  membership: DAOMembership;
  earlyExecution: boolean;
  voteReplacement: boolean;
  multisigWallets: MultisigWalletField[];
  multisigMinimumApprovals: number;

  votingType: 'onChain' | 'gasless';
  executionExpirationMinutes: string;
  executionExpirationHours: string;
  executionExpirationDays: string;
  committee: MultisigWalletField[];
  committeeMinimumApproval: string;
};